complete order and payment function, refactor interfaces (use tailwindcss, remove css and bootstrap)

+++ be/src/controllers/cart.controller.js
@@ -8,18 +8,21 @@ export default class CartController {
     addProduct = async (req, res) => {
         try {
             const { user } = req;
-            const { product, quantity } = req.body;
+            const { product, quantity, product_option } = req.body;
 
-            const cart = await this.getOne({
+            let cart = await new CartService().getOne({
                 where: { user_id: user.id },
-                include: [{ model: db.cart_item, include: [db.product] }]
+                include: [{ model: db.cart_item, include: [{ model: db.product, include: [db.product_option] }, { model: db.product_option }] }]
             });
-            const cartItem = await new CartItemService().createCartItem(cart.id, product.id, quantity);
+            const cartItem = await new CartItemService().createCartItem(cart.id, product.id, quantity, product_option.id);
 
             cart.cart_items.push(cartItem);
 
-            await new CartService().update(cart);
-            console.log('test')
+            cart = await new CartService().getOne({
+                where: { user_id: user.id },
+                include: [{ model: db.cart_item, include: [{ model: db.product, include: [db.product_option] }, { model: db.product_option }] }]
+            });
+
             return res.status(200).json(cart);
         } catch (error) {
             console.error(error);
@@ -27,13 +30,13 @@ export default class CartController {
         }
     }
 
-    getByuser_id = async (req, res) => {
+    getByUserId = async (req, res) => {
         try {
-            const { user } = req;
-            let cart = await new CartService().getCartByuser_id(user.id);
+            const user = req.user;
+            let cart = await new CartService().getByUserId(user.id);
 
             if (!cart) {
-                cart = await new CartService().createCart(user.id);
+                cart = await new CartService().create({ user_id: user.id });
             }
 
             if (!cart.cart_items) {
@@ -47,18 +50,62 @@ export default class CartController {
         }
     }
 
+    updateCartItem = async (req, res) => {
+        try {
+            const { user } = req;
+            const { quantity, product_option } = req.body;
+            console.log(req.body)
+
+            await new CartItemService().update({
+                id: req.params.id,
+                quantity: quantity,
+                product_option_id: product_option.id
+            });
+
+            const cart = await new CartService().getOne({
+                where: { user_id: user.id },
+                include: [{ model: db.cart_item, include: [{ model: db.product, include: [db.product_option] }, { model: db.product_option }] }]
+            });
+
+            return res.status(200).json(cart);
+        }
+        catch (error) {
+            console.error(error);
+            return res.status(500).json({ message: "Internal server error" });
+        }
+    }
+
+    deleteCartItem = async (req, res) => {
+        try {
+            const { user } = req;
+
+            await new CartItemService().delete(req.params.id);
+
+            const cart = await new CartService().getOne({
+                where: { user_id: user.id },
+                include: [{ model: db.cart_item, include: [{ model: db.product, include: [db.product_option] }, { model: db.product_option }] }]
+            });
+
+            return res.status(200).json(cart);
+        }
+        catch (error) {
+            console.error(error);
+
+            return res.status(500).json({ message: "Internal server error" });
+        }
+    }
+
     sync = async (req, res) => {
         try {
             const { user } = req;
-            const cart = await this.getCartByuser_id(user.id);
+            const cart = await new CartService().getByUserId(user.id);
             const { cart_items } = req.body;
 
             for (const item of cart_items) {
-                const cartItem = cart.cart_items.find(user_item => user_item.product.id === item.product.id);
-
+                const cartItem = cart.cart_items?.find(user_item => user_item.product.id === item.product.id);
                 if (cartItem) {
                     cartItem.quantity += item.quantity;
-                    await new CartItemService().updateCartItemQuantity(cartItem);
+                    await new CartItemService().updateCartItem(cartItem);
                 } else {
                     await new CartItemService().createCartItem(cart.id, item.product.id, item.quantity);
                 }
@@ -69,10 +116,6 @@ export default class CartController {
             console.error(error);
             return res.status(500).json({ message: "Internal server error" });
         }
-    }
-
-    getCartByuser_id = async (user_id) => {
-        return new CartService().getCartByuser_id(user_id);
     }
 
     async getAll(req, res) {
@@ -120,6 +163,14 @@ export default class CartController {
 
     async delete(req, res) {
         try {
+            const { user } = req;
+            let cart = await new CartService().getByUserId(user.id);
+            if (!cart) {
+                return res.status(404).json({ message: "Not found" });
+            }
+            if (cart.id !== parseInt(req.params.id)) {
+                return res.status(403).json({ message: "Forbidden" });
+            }
             await new CartService().delete(req.params.id);
             return res.status(204).json();
         } catch (error) {
